+++ src/PhpWord/Element/Footnote.php
@@ -50,7 +50,7 @@ class Footnote extends AbstractContainer
      */
     public function __construct($paragraphStyle = null)
     {
-        $this->paragraphStyle = $this->setStyle(new Paragraph(), $paragraphStyle);
+        $this->paragraphStyle = $this->setNewStyle(new Paragraph(), $paragraphStyle);
         $this->setDocPart($this->container);
     }