+++ src/PhpWord/Element/ListItemRun.php
@@ -59,9 +59,9 @@ class ListItemRun extends TextRun
         if (!is_null($listStyle) && is_string($listStyle)) {
             $this->style = new ListItemStyle($listStyle);
         } else {
-            $this->style = $this->setStyle(new ListItemStyle(), $listStyle, true);
+            $this->style = $this->setNewStyle(new ListItemStyle(), $listStyle, true);
         }
-        $this->paragraphStyle = $this->setStyle(new Paragraph(), $paragraphStyle);
+        $this->paragraphStyle = $this->setNewStyle(new Paragraph(), $paragraphStyle);
     }
 
     /**