+++ src/PhpWord/Element/Section.php
@@ -18,7 +18,7 @@
 namespace PhpOffice\PhpWord\Element;
 
 use PhpOffice\PhpWord\Exception\Exception;
-use PhpOffice\PhpWord\Style\Section as SectionSettings;
+use PhpOffice\PhpWord\Style\Section as SectionStyle;
 
 /**
  * Section
@@ -31,11 +31,11 @@ class Section extends AbstractContainer
     protected $container = 'Section';
 
     /**
-     * Section settings
+     * Section style
      *
      * @var \PhpOffice\PhpWord\Style\Section
      */
-    private $settings;
+    private $style;
 
     /**
      * Section headers, indexed from 1, not zero
@@ -55,41 +55,36 @@ class Section extends AbstractContainer
      * Create new instance
      *
      * @param int $sectionCount
-     * @param array $settings
+     * @param array $style
      */
-    public function __construct($sectionCount, $settings = null)
+    public function __construct($sectionCount, $style = null)
     {
         $this->sectionId = $sectionCount;
         $this->setDocPart($this->container, $this->sectionId);
-        $this->settings = new SectionSettings();
-        $this->setSettings($settings);
+        $this->style = new SectionStyle();
+        $this->setStyle($style);
     }
 
     /**
-     * Set section settings
+     * Set section style
      *
-     * @param array $settings
+     * @param array $style
      */
-    public function setSettings($settings = null)
+    public function setStyle($style = null)
     {
-        if (!is_null($settings) && is_array($settings)) {
-            foreach ($settings as $key => $value) {
-                if (is_null($value)) {
-                    continue;
-                }
-                $this->settings->setSettingValue($key, $value);
-            }
+        if (!is_null($style) && is_array($style)) {
+            $this->style->setStyleByArray($style);
         }
     }
 
     /**
-     * Get Section Settings
+     * Get section style
      *
      * @return \PhpOffice\PhpWord\Style\Section
      */
-    public function getSettings()
+    public function getStyle()
     {
-        return $this->settings;
+        return $this->style;
     }
 
     /**
@@ -182,6 +177,30 @@ class Section extends AbstractContainer
             throw new Exception('Invalid header/footer type.');
         }
 
+    }
+
+    /**
+     * Set section style
+     *
+     * @param array $settings
+     * @deprecated 0.12.0
+     * @codeCoverageIgnore
+     */
+    public function setSettings($settings = null)
+    {
+        $this->setStyle($settings);
+    }
+
+    /**
+     * Get section style
+     *
+     * @return \PhpOffice\PhpWord\Style\Section
+     * @deprecated 0.12.0
+     * @codeCoverageIgnore
+     */
+    public function getSettings()
+    {
+        return $this->getStyle();
     }
 
     /**